refactor: change deploy structure
make possible to use "python -m deploy"
