+++ deploy/run.py
@@ -1,43 +1,46 @@
 """
-Automatic deployment with Ansible and ansible-runner
+Package for deployment
 """
-import asyncio
 import logging
-import logging.config
-import os
 
-from concurrent.futures import ThreadPoolExecutor
-
-from deploy import perform_deployment
-from deploy.deploy_const import PROJECT_ROOT_PATH
+from ansible import AnsibleExecutor
+from deploy.deploy_const import DEVELOPMENT_MODE, PRODUCTION_MODE
+from deploy.jobs import *
+from deploy.utils import create_directory
 from settings import config
 
+logger = logging.getLogger('ansible_deploy')
 
-def configure_deploy_logging_locally(logger_file: str):
+
+async def perform_deployment(deploy_mode: str, local_output_dir: str):
     """
-    Configures deploy logging file locally
+    Deployment entry point
 
     Args:
-        logger_file: path of the logger file
+        deploy_mode: mode of deployment
+        local_output_dir: path to an existing local directory to be used:
+                              - as ansible-runner's private data directory
     """
-    dir_path = os.path.dirname(logger_file)
-    try:
-        os.makedirs(dir_path)
-        print(f"Directory {dir_path} successfully created")
-    except FileExistsError:
-        print(f"Directory {dir_path} already exist")
+    create_directory(local_output_dir)
 
-    deploy_log_config = config.logging_ansible_deploy
-    if not deploy_log_config['handlers']['service_file']['filename']:
-        deploy_log_config['handlers']['service_file']['filename'] = logger_file
-    logging.config.dictConfig(config=deploy_log_config)
+    assert deploy_mode in (PRODUCTION_MODE, DEVELOPMENT_MODE), \
+        f"Only two modes of deployment is allowed: '{DEVELOPMENT_MODE}' and '{PRODUCTION_MODE}'"
 
+    target_host = config.server.ip
+    logger.info("Initiate '%s' mode of deployment on '%s' host", deploy_mode, target_host)
 
-if __name__ == '__main__':
-    configure_deploy_logging_locally(logger_file=os.path.join(PROJECT_ROOT_PATH.joinpath('tmp/'), 'ansible-deploy.log'))
+    ansible_executor = AnsibleExecutor(host_pattern=target_host,
+                                       private_data_dir=local_output_dir,
+                                       verbosity=config.ansible.verbosity)
+    logger.debug("Successfully initiate instance of '%s' class", AnsibleExecutor.__name__)
 
-    asyncio.new_event_loop().set_default_executor(ThreadPoolExecutor(max_workers=2))
-    asyncio.run(
-        perform_deployment(deploy_mode=config.deploy.mode,
-                           local_output_dir=PROJECT_ROOT_PATH.joinpath('tmp/ansible'))
-    )
+    await echo_host(ansible=ansible_executor, need_gather_facts=deploy_mode == PRODUCTION_MODE)
+
+    logger.debug("Preparing the project for deployment")
+    await make_preparation(ansible=ansible_executor)
+
+    logger.debug("Docker installation")
+    await install_docker(ansible=ansible_executor)
+
+    logger.debug("Configure Nginx as a web-server")
+    await configure_nginx(ansible=ansible_executor)
